Add int conversion back. Should we round instead?

+++ posenet/decode_multi.py
@@ -134,7 +134,8 @@ def decode_multiple_poses(
     displacements_bwd = displacements_bwd.reshape(height, width, 2, -1).swapaxes(2, 3)
 
     for root_score, root_id, root_coord in scored_parts:
-        root_image_coords = root_coord * output_stride + offsets[root_coord[0], root_coord[1], root_id]
+        root_image_coords = root_coord * output_stride + offsets[
+            root_coord[0], root_coord[1], root_id].astype(np.int32)
 
         if within_nms_radius_fast(
                 pose_keypoint_coords[:pose_count, root_id, :], squared_nms_radius, root_image_coords):
